requirements.txt
#fritzconnection==1.0.1
fritzconnection>=0.8.4
future==0.18.2
matplotlib==3.5.2
pandas==1.3.2
seaborn==0.11.2
numpy==1.26.4
paho-mqtt<2.0.0
schedule<=1.2.2